+++ packages/interaction-kit/CHANGELOG.md
@@ -16,7 +16,7 @@
   - discord-edge-runner@1.0.0
   - discord-verify@1.0.2
   - discord-request@0.0.6
-  - discord-api@0.0.4
+  - discord-api-methods@0.0.4
 
 ## 0.2.2